Nats stream output

+++ src/Argon.Api/Grains.Interfaces/ITestGrain.cs
@@ -4,7 +4,7 @@ namespace Argon.Api.Grains.Interfaces;
 [DataContract, MemoryPackable(GenerateType.VersionTolerant), MessagePackObject, Serializable, GenerateSerializer, Alias(nameof(SomeInput))]
 public sealed partial record SomeInput(
     [property: DataMember(Order = 0), MemoryPackOrder(0), Key(0), Id(0)]
-    int a,
+    long a,
     [property: DataMember(Order = 1), MemoryPackOrder(1), Key(1), Id(1)]
     string b);
